new: Participant
I commited this to a new branch because someone had already thought of doing this in the Storage package.

However, I disagree, the Storage should be really dumb and the transaction logic contained to the Transaction package, if possible.

This way a StorageServer can extend the base class Participant and choose which kind of IStorage to use.